add readme / fix bug

+++ core/follow/service.go
@@ -2,8 +2,11 @@ package follow
 
 import (
 	"context"
+	"log"
 
 	"meh/core/user"
+
+	"github.com/pkg/errors"
 )
 
 type Service struct {
@@ -17,15 +20,35 @@ func NewService(repo Repository) *Service {
 }
 
 func (s *Service) Follow(ctx context.Context, from, to user.ID) error {
+	if from == to {
+		return errors.New("cannot follow same id")
+	}
+
+	found, err := s.repo.Exists(ctx, from, to)
+	if err != nil {
+		return errors.Wrapf(err, "failed to check follow from=%v to=%v", from, to)
+	}
+
+	if found {
+		log.Printf("follow is ignored because alreday followed from=%v to=%v", from, to)
+		return nil
+	}
+
 	return s.repo.Create(ctx, from, to)
 }
 
 func (s *Service) Remove(ctx context.Context, from, to user.ID) error {
-	return s.repo.Delete(ctx, from, to)
-}
+	found, err := s.repo.Exists(ctx, from, to)
+	if err != nil {
+		return errors.Wrapf(err, "failed to check follow from=%v to=%v", from, to)
+	}
 
-func (s *Service) Exists(ctx context.Context, from, to user.ID) (bool, error) {
-	return s.repo.Exists(ctx, from, to)
+	if !found {
+		log.Printf("remove is ignored because not followed from=%v to=%v", from, to)
+		return nil
+	}
+
+	return s.repo.Delete(ctx, from, to)
 }
 
 func (s *Service) ListFollowers(ctx context.Context, userID user.ID) ([]user.ID, error) {
